% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{mpt.feedback.no_score}
\alias{mpt.feedback.no_score}
\title{MPT feedback (no score)}
\usage{
mpt.feedback.no_score(dict = mpt::mpt_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
demo_mpt(feedback = mpt.feedback.no_score())}
}
