% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mpt.R
\name{theme_mpt}
\alias{theme_mpt}
\title{Establece algunas características estéticas de los gráficos
de Mercados para Todos}
\usage{
theme_mpt()
}
\value{
No devuelve valores
}
\description{
Establece algunas características estéticas de los gráficos
de Mercados para Todos
}
\details{
Define el theme con menos rayas horizontales y verticales.
Posición de la leyenda.
Pone el título en bold.

#'
}
\examples{
ggplot() + theme_mpt() -> Establece los parámetros al theme

}
