% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt.R
\name{mpt}
\alias{mpt}
\title{MPT}
\usage{
mpt(num_items = 30L, take_training = TRUE, label = "MPT",
  feedback = mpt.feedback.no_score(),
  audio_dir = "https://media.gold-msi.org/test_materials/MPT/v1-1/audio",
  training_dir = "https://media.gold-msi.org/test_materials/MPT/training",
  next_item.criterion = "bOpt", next_item.estimator = "BM",
  next_item.prior_dist = "norm", next_item.prior_par = c(0, 1),
  final_ability.estimator = "WL", constrain_answers = FALSE,
  dict = mpt::mpt_dict)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{take_training}{(Logical scalar) Whether to include the training phase.}

\item{label}{(Character scalar) Label to give the MPT results in the output file.}

\item{feedback}{Defines the feedback to give the participant
at the end of the test. By default no feedback is given.
This can be a timeline segment (as created by \code{\link[psychTestR]{new_timeline}}),
a test element (as created by e.g. \code{\link[psychTestR]{page}}),
or a list of test elements.
The following built-in choices are available
(see function-level documentation for details):
\itemize{
\item \code{\link{mpt.feedback.no_score}}
\item \code{\link[psychTestRCAT]{cat.feedback.graph}}
\item \code{\link[psychTestRCAT]{cat.feedback.irt}}
\item \code{\link[psychTestRCAT]{cat.feedback.iq}}
}}

\item{audio_dir}{(Character scalar) File path to the directory
hosting the audio (typically a publicly accessible web directory).}

\item{training_dir}{(Character scalar) File path to the directory
hosting the training items' audio (typically a publicly accessible web directory).}

\item{next_item.criterion}{(Character scalar)
Criterion for selecting successive items in the adaptive test.
See the \code{criterion} argument in \code{\link[catR]{nextItem}} for possible values.
\code{"bOpt"} corresponds to the setting used in the original MPT paper.}

\item{next_item.estimator}{(Character scalar)
Ability estimation method used for selecting successive items in the adaptive test.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
\code{"BM"}, Bayes modal,
corresponds to the setting used in the original MPT paper.
\code{"WL"}, weighted likelihood,
corresponds to the default setting used in versions <= 0.2.0 of this package.}

\item{next_item.prior_dist}{(Character scalar)
The type of prior distribution to use when calculating ability estimates
for item selection.
Ignored if \code{next_item.estimator} is not a Bayesian method.
Defaults to \code{"norm"} for a normal distribution.
See the \code{priorDist} argument in \code{\link[catR]{thetaEst}} for possible values.}

\item{next_item.prior_par}{(Numeric vector, length 2)
Parameters for the prior distribution;
see the \code{priorPar} argument in \code{\link[catR]{thetaEst}} for details.
Ignored if \code{next_item.estimator} is not a Bayesian method.
The default is \code{c(0, 1)}.}

\item{final_ability.estimator}{Estimation method used for the final ability estimate.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
The default is \code{"WL"}, weighted likelihood,
which corresponds to the setting used in the original MPT paper.
If a Bayesian method is chosen, its prior distribution will be defined
by the \code{next_item.prior_dist} and \code{next_item.prior_par} arguments.}

\item{constrain_answers}{(Logical scalar)
If \code{TRUE}, then item selection will be constrained so that the
correct answers are distributed as evenly as possible over the course of the test.
We recommend leaving this option disabled.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a MPT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MPT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the MPT, consider using \code{\link{demo_mpt}()}.
For a standalone implementation of the MPT,
consider using \code{\link{standalone_mpt}()}.
}
\note{
Versions <= 0.2.0 of this package experimented with weighted likelihood
ability estimation for item selection.
However, subsequent versions of the package revert to Bayes modal
ability estimation for item selection, consistent with the original MPT paper.
}
